[Refactor] Implement book log feature
